fix(data): firestoreは`uint`をサポートしていない

+++ internal/infrastructure/firestore/entity/place.go
@@ -3,11 +3,12 @@ package entity
 import "poroto.app/poroto/planner/internal/domain/models"
 
 type PlaceEntity struct {
-	Name                  string            `firestore:"name"`
-	Location              GeoLocationEntity `firestore:"location"`
-	Thumbnail             *string           `firestore:"thumbnail"`
-	Photos                []string          `firestore:"photos"`
-	EstimatedStayDuration uint              `firestore:"estimated_stay_duration"`
+	Name      string            `firestore:"name"`
+	Location  GeoLocationEntity `firestore:"location"`
+	Thumbnail *string           `firestore:"thumbnail"`
+	Photos    []string          `firestore:"photos"`
+	// MEMO: Firestoreではuintをサポートしていないため，intにしている
+	EstimatedStayDuration int `firestore:"estimated_stay_duration"`
 }
 
 func ToPlaceEntity(place models.Place) PlaceEntity {
@@ -16,7 +17,7 @@ func ToPlaceEntity(place models.Place) PlaceEntity {
 		Location:              ToGeoLocationEntity(place.Location),
 		Thumbnail:             place.Thumbnail,
 		Photos:                place.Photos,
-		EstimatedStayDuration: place.EstimatedStayDuration,
+		EstimatedStayDuration: int(place.EstimatedStayDuration),
 	}
 }
 
@@ -26,6 +27,6 @@ func FromPlaceEntity(entity PlaceEntity) models.Place {
 		Location:              FromGeoLocationEntity(entity.Location),
 		Thumbnail:             entity.Thumbnail,
 		Photos:                entity.Photos,
-		EstimatedStayDuration: entity.EstimatedStayDuration,
+		EstimatedStayDuration: uint(entity.EstimatedStayDuration),
 	}
 }
